requirements.txt
RPi.GPIO
numpy
scipy
sounddevice
openai>=1.50.0
python-dotenv 
websocket-client